+++ pyproject.toml
@@ -37,6 +37,7 @@ ndlib = "^5.1.1"
 networkx = "^3.2.1"
 pandas = ">=2.0.0"
 tqdm = "^4.66.2"
+pooch = "^1.8.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.2"
