+++ docs/migration.md
@@ -1,0 +1,266 @@
+# Automata Migration Guide
+
+## Backwards-incompatible changes from v7 to v8
+
+### Dependency Changes
+
+Python 3.7 support has been dropped. Please upgrade to Python 3.8 or later to
+use Automata v8.
+
+Diagrams are no longer being generated using `pydot`; this dependency has been
+dropped in favor of using the `visual` optional dependency, which will install
+`pygraphviz` and `coloraide` used for generating figures. You should install
+this optional dependency if you wish to generate figures. This change was to
+allow for native support for displaying finite automaton in Jupyter notebooks.
+The style of the diagrams has been lifted from the [visual automata] package,
+so you should take a look at the diagrams generated and see if they are still
+satisfactory.
+
+[visual automata]: https://pypi.org/project/visual-automata/
+
+Other new dependencies have been added, but these will be installed automatically
+along with v8 of the package.
+
+### Greater Support for Partial DFAs
+
+There is now greater support for partial DFAs, which included changing the
+`DFA.from_nfa()` function to return a partial DFA instead of a complete one.
+To obtain a complete DFA, you must now call `DFA.from_nfa().to_complete(trap_state_name)`,
+where `trap_state_name` will be used as the name for a trap state if one needs to
+be added.
+
+### Type Hints
+
+Type hints have now been added, meaning that code which previously called functions
+with incorrect types may not have been flagged. See output from your typechecker
+for more information.
+
+### NFA.from_regex default input symbols
+The default set of input symbols for `NFA.from_regex` was changed to all ascii letters and digits.
+If needing to use a specific set of input symbols, use the `input_symbols` parameter.
+
+## Backwards-incompatible changes from v6 to v7
+
+### Immutable instances
+
+All Automaton instances are now fully immutable to protect against common
+pitfalls, such as mutating an automaton to an invalid state after it's already
+been validated.
+
+This means that if you wish to make a change to an automaton instance, you must
+retrieve its attributes as a dictionary (using the new `input_parameters`
+property), make your desired change, then pass those parameters to the relevant
+constructor. For example:
+
+```python
+from automata.fa.dfa import DFA
+
+dfa1 = DFA(
+    states={'q0', 'q1', 'q2'},
+    input_symbols={'0', '1'},
+    transitions={
+        'q0': {'0': 'q0', '1': 'q1'},
+        'q1': {'0': 'q0', '1': 'q2'},
+        'q2': {'0': 'q2', '1': 'q1'}
+    },
+    initial_state='q0',
+    final_states={'q1'}
+)
+# You can still copy an automaton just fine
+dfa2 = dfa.copy()
+# If you want to make a change, you must create a new instance; please note
+# that dfa2.input_parameters is always a deep copy of the input parameters for
+# dfa2 (in other words, mutating dfa2.input_parameters will not actually mutate
+# dfa2)
+params = dfa2.input_parameters
+params['final_states'] = {'q2'}
+dfa3 = DFA(**params)
+```
+
+### Renamed Regex Module
+
+The `automata.base.regex` module has been renamed to `automata.regex.regex`
+alongside the other regular expression-related modules.
+
+### DFA.minify() defaults
+
+The default value of the `retain_names` parameter for `DFA.minify()` has been
+corrected from `True` to `False`; the API documentation has always stated that
+the default value _should_ be `False`, however the default value in the code was
+actually `True`; therefore, the code has been updated to match the documentation
+(#59)
+ - Since this code correction may break existing developer code, this is labeled
[…]
