+++ docs/api/class-automaton.md
@@ -1,3 +1,0 @@
-# class Automaton(metaclass=ABCMeta)
-
-::: cynetdiff.utils.networkx_to_ic_model
