+++ docs/api/models.md
@@ -1,0 +1,1 @@
+::: cynetdiff.models
