### 2024-03-11 bb88461
First version of act prob utility functions
