### 2024-03-11 a36879c
Different constructor
