### 2024-03-11 b80971e
Added weighting scheme functions
