### 2024-03-11 5316526
simplified function
