### 2024-03-11 767e759
Added normalization check
