### 2024-03-11 4f789c8
Simplify tets function
