### 2024-03-12 f60391e
Added contributing guide
