### 2024-03-13 af47442
Update CITATION.cff
