+++ CITATION.cff
@@ -3,7 +3,13 @@ message: "If you use this software, please cite it as below."
 authors:
 - family-names: Robson
   given-names: Eliot W.
-orcid: "https://orcid.org/0000-0002-1476-6715"
-title: "CyNetDiff: v0.1.5"
+  orcid: https://orcid.org/0000-0002-1476-6715
+- family-names: Umrawal
+  given-names: Abhishek K.
+  orcid: https://orcid.org/0000-0003-4460-7499
+title: "CyNetDiff"
+identifiers:
+  - type: doi
+    value: 10.5281/zenodo.10801504
 version: v0.1.5
 date-released: 2024-03-10
