+++ src/cynetdiff/models.cpp
@@ -3093,6 +3093,9 @@ static CYTHON_INLINE PyObject* __Pyx_PyInt_From_unsigned_int(unsigned int value)
 /* CIntFromPy.proto */
 static CYTHON_INLINE unsigned int __Pyx_PyInt_As_unsigned_int(PyObject *);
 
+/* CIntFromPy.proto */
+static CYTHON_INLINE size_t __Pyx_PyInt_As_size_t(PyObject *);
+
 /* CIntToPy.proto */
 static CYTHON_INLINE PyObject* __Pyx_PyInt_From_long(long value);
 
@@ -23787,8 +23790,8 @@ static PyObject *__pyx_pf_9cynetdiff_6models_20LinearThresholdModel_2set_seeds(s
  * 
  * 
  *     cpdef void reassign_threshold(self):             # <<<<<<<<<<<<<<
- *         for i in range(self.threshold.size()):
- *             self.threshold[i] = next_rand()
+ *         for i in range(self.thresholds.size()):
+ *             self.thresholds[i] = next_rand()
  */
 
 static PyObject *__pyx_pw_9cynetdiff_6models_20LinearThresholdModel_5reassign_threshold(PyObject *__pyx_v_self, 
@@ -23799,16 +23802,17 @@ PyObject *__pyx_args, PyObject *__pyx_kwds
 #endif
 ); /*proto*/
 static void __pyx_f_9cynetdiff_6models_20LinearThresholdModel_reassign_threshold(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *__pyx_v_self, int __pyx_skip_dispatch) {
-  PyObject *__pyx_v_i = NULL;
+  std::vector<float> ::size_type __pyx_v_i;
   __Pyx_RefNannyDeclarations
   PyObject *__pyx_t_1 = NULL;
   PyObject *__pyx_t_2 = NULL;
   PyObject *__pyx_t_3 = NULL;
   PyObject *__pyx_t_4 = NULL;
   int __pyx_t_5;
-  Py_ssize_t __pyx_t_6;
-  PyObject *(*__pyx_t_7)(PyObject *);
-  double __pyx_t_8;
+  std::vector<float> ::size_type __pyx_t_6;
+  std::vector<float> ::size_type __pyx_t_7;
+  std::vector<float> ::size_type __pyx_t_8;
+  double __pyx_t_9;
   int __pyx_lineno = 0;
   const char *__pyx_filename = NULL;
   int __pyx_clineno = 0;
@@ -23868,128 +23872,32 @@ static void __pyx_f_9cynetdiff_6models_20LinearThresholdModel_reassign_threshold
   /* "cynetdiff/models.pyx":180
  * 
  *     cpdef void reassign_threshold(self):
- *         for i in range(self.threshold.size()):             # <<<<<<<<<<<<<<
- *             self.threshold[i] = next_rand()
+ *         for i in range(self.thresholds.size()):             # <<<<<<<<<<<<<<
+ *             self.thresholds[i] = next_rand()
  * 
  */
-  __pyx_t_2 = __Pyx_PyObject_GetAttrStr(((PyObject *)__pyx_v_self), __pyx_n_s_threshold); if (unlikely(!__pyx_t_2)) __PYX_ERR(0, 180, __pyx_L1_error)
-  __Pyx_GOTREF(__pyx_t_2);
-  __pyx_t_3 = __Pyx_PyObject_GetAttrStr(__pyx_t_2, __pyx_n_s_size); if (unlikely(!__pyx_t_3)) __PYX_ERR(0, 180, __pyx_L1_error)
[…]
