+++ src/cynetdiff/models.cpp
@@ -1737,9 +1737,9 @@ struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel {
   struct __pyx_obj_9cynetdiff_6models_DiffusionModel __pyx_base;
   __Pyx_memviewslice starts;
   __Pyx_memviewslice edges;
-  float threshold;
-  __Pyx_memviewslice edge_thresholds;
-  __Pyx_memviewslice edge_probabilities;
+  float activation_prob;
+  __Pyx_memviewslice activation_probs;
+  __Pyx_memviewslice _edge_probabilities;
   std::deque<unsigned int>  work_deque;
   std::unordered_set<unsigned int>  seen_set;
   std::unordered_set<unsigned int>  original_seeds;
@@ -1782,7 +1782,7 @@ struct __pyx_obj_9cynetdiff_6models___pyx_scope_struct__get_newly_activated_node
 };
 
 
-/* "cynetdiff/models.pyx":188
+/* "cynetdiff/models.pyx":189
  *         self.seen_set.insert(self.original_seeds.begin(), self.original_seeds.end())
  * 
  *     def get_newly_activated_nodes(self):             # <<<<<<<<<<<<<<
@@ -3488,7 +3488,6 @@ static const char __pyx_k_isenabled[] = "isenabled";
 static const char __pyx_k_pyx_state[] = "__pyx_state";
 static const char __pyx_k_reduce_ex[] = "__reduce_ex__";
 static const char __pyx_k_set_seeds[] = "set_seeds";
-static const char __pyx_k_threshold[] = "threshold";
 static const char __pyx_k_IndexError[] = "IndexError";
 static const char __pyx_k_ValueError[] = "ValueError";
 static const char __pyx_k_pyx_result[] = "__pyx_result";
@@ -3512,18 +3511,19 @@ static const char __pyx_k_reduce_cython[] = "__reduce_cython__";
 static const char __pyx_k_AssertionError[] = "AssertionError";
 static const char __pyx_k_DiffusionModel[] = "DiffusionModel";
 static const char __pyx_k_View_MemoryView[] = "View.MemoryView";
+static const char __pyx_k_activation_prob[] = "activation_prob";
 static const char __pyx_k_allocate_buffer[] = "allocate_buffer";
 static const char __pyx_k_collections_abc[] = "collections.abc";
 static const char __pyx_k_dtype_is_object[] = "dtype_is_object";
-static const char __pyx_k_edge_thresholds[] = "edge_thresholds";
 static const char __pyx_k_pyx_PickleError[] = "__pyx_PickleError";
 static const char __pyx_k_setstate_cython[] = "__setstate_cython__";
+static const char __pyx_k_activation_probs[] = "activation_probs";
 static const char __pyx_k_cynetdiff_models[] = "cynetdiff.models";
 static const char __pyx_k_successor_starts[] = "successor_starts";
 static const char __pyx_k_pyx_unpickle_Enum[] = "__pyx_unpickle_Enum";
 static const char __pyx_k_asyncio_coroutines[] = "asyncio.coroutines";
 static const char __pyx_k_cline_in_traceback[] = "cline_in_traceback";
-static const char __pyx_k_edge_probabilities[] = "edge_probabilities";
[…]
