### 2024-03-13 5163e33
Made benchmark dependencies optional
