+++ pyproject.toml
@@ -74,11 +74,6 @@ build-backend = "poetry.core.masonry.api"
 script = "build_package.py"
 generate-setup-file = false
 
-[tool.flake8]
-count = true
-statistics = true
-# Black compatibility
-max-line-length = 88
 
 [tool.mypy]
 python_version = "3.10"
