+++ tests/test_lt_model.py
@@ -215,7 +215,7 @@ def test_specific_model(directed: bool, nondefault_influence: bool) -> None:
     total_num = sum(map(len, activated_nodes_levels))
 
     # Reassigning the threshold should change the total number of activated nodes
-    model.reassign_threshold()
+    model.reassign_thresholds()
     model.advance_until_completion()
 
     # TODO this test passes with high-enough probability. Refactor to avoid a possible
