+++ src/cynetdiff/models.cpp
@@ -1782,7 +1782,7 @@ struct __pyx_obj_9cynetdiff_6models___pyx_scope_struct__get_newly_activated_node
 };
 
 
-/* "cynetdiff/models.pyx":189
+/* "cynetdiff/models.pyx":193
  *         self.seen_set.insert(self.original_seeds.begin(), self.original_seeds.end())
  * 
  *     def get_newly_activated_nodes(self):             # <<<<<<<<<<<<<<
@@ -1917,7 +1917,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
 
 struct __pyx_vtabstruct_9cynetdiff_6models_LinearThresholdModel {
   struct __pyx_vtabstruct_9cynetdiff_6models_DiffusionModel __pyx_base;
-  void (*reassign_threshold)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, int __pyx_skip_dispatch);
+  void (*reassign_thresholds)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, int __pyx_skip_dispatch);
   int (*_LinearThresholdModel__activation_succeeds)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, unsigned int, std::unordered_set<unsigned int>  const &);
   int (*_LinearThresholdModel__advance_model)(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *, std::deque<unsigned int>  &, std::unordered_set<unsigned int>  &);
 };
@@ -3233,7 +3233,7 @@ static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCasc
 static void __pyx_f_9cynetdiff_6models_23IndependentCascadeModel_advance_until_completion(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
 static void __pyx_f_9cynetdiff_6models_23IndependentCascadeModel_advance_model(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
 static int __pyx_f_9cynetdiff_6models_23IndependentCascadeModel__IndependentCascadeModel__advance_model(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self, std::deque<unsigned int>  &__pyx_v_work_deque, std::unordered_set<unsigned int>  &__pyx_v_seen_set); /* proto*/
-static void __pyx_f_9cynetdiff_6models_20LinearThresholdModel_reassign_threshold(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
+static void __pyx_f_9cynetdiff_6models_20LinearThresholdModel_reassign_thresholds(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
 static void __pyx_f_9cynetdiff_6models_20LinearThresholdModel_reset_model(struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel *__pyx_v_self, int __pyx_skip_dispatch); /* proto*/
[…]
