+++ docs/people.md
@@ -1,13 +1,7 @@
 ## Maintainers
 
-- [caleb531](https://github.com/caleb531)
 - [eliotwrobson](https://github.com/eliotwrobson)
 
 ## Contributors
 
-- [lewiuberg](https://github.com/lewiuberg)
-- [YtvwlD](https://github.com/YtvwlD)
-- [dengl11](https://github.com/dengl11)
-- [Tagl](https://github.com/Tagl)
-- [CamiloMartinezM](https://github.com/CamiloMartinezM)
-- [abhinavsinha-adrino](https://github.com/abhinavsinha-adrino)
+- [abhishekumrawal](https://github.com/abhishekumrawal)
