+++ docs/contributing.md
@@ -1,0 +1,3 @@
+--8<-- "CONTRIBUTING.md"
+
+<!-- TODO figure out how to get link to code of conduct working ok. -->
