+++ .github/workflows/lint.yml
@@ -14,14 +14,14 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.12"]
-        poetry-version: ["1.7.1"]
+        poetry-version: ["1.8.1"]
         os: [ubuntu-latest]
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
         with:
           submodules: recursive
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
