+++ .github/workflows/tests.yml
@@ -15,15 +15,14 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        # "3.8", "3.9", "3.10",
-        python-version: ["3.11"]
-        poetry-version: ["1.7.1"]
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        poetry-version: ["1.8.1"]
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v3
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
 
