+++ .github/workflows/release.yml
@@ -1,174 +1,0 @@
-name: Release Python Package
-on:
-  workflow_dispatch:
-
-jobs:
-  build-linux:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["cp37-cp37m", "cp38-cp38", "cp39-cp39", "cp310-cp310", "cp311-cp311"]
-    container: quay.io/pypa/manylinux_2_28_x86_64
-    steps:
-      - uses: actions/checkout@v3
-      - name: Build binary wheel ${{ matrix.python-version }}
-        run: |
-          /opt/python/${{ matrix.python-version }}/bin/python -m pip install -r build_requirements.txt
-          /opt/python/${{ matrix.python-version }}/bin/python -m pip wheel . -w wheelhouse/
-          auditwheel repair wheelhouse/leafy*.whl -w dist/
-
-      - name: Archive dist artifacts
-        uses: actions/upload-artifact@v2
-        with:
-          name: dist-linux-${{ matrix.python-version }}
-          path: dist
-
-  build-macos-arm:
-    runs-on: macos-13-xlarge
-    strategy:
-      matrix:
-        python-version: ["3.11"]
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }} x64
-        uses: actions/setup-python@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-          architecture: x64
-
-      - name: build package
-        run: |
-          python -m pip install -r build_requirements.txt
-          python setup.py bdist_wheel
-
-      - name: Archive dist artifacts
-        uses: actions/upload-artifact@v2
-        with:
-          name: dist-macos-arm-${{ matrix.python-version }}
-          path: dist
-
-  build-macos-intel:
-    runs-on: macos-latest
-    strategy:
-      matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11"]
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }} x64
-        uses: actions/setup-python@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-          architecture: x64
-
-      - name: build package
-        run: |
-          python -m pip install -r build_requirements.txt
-          python setup.py bdist_wheel
-
-      - name: Archive dist artifacts
-        uses: actions/upload-artifact@v2
-        with:
-          name: dist-macos-${{ matrix.python-version }}
-          path: dist
-
-  build-windows:
-    runs-on: windows-latest
-    strategy:
-      matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11"]
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }} x64
-        uses: actions/setup-python@v3
-        with:
-          python-version: ${{ matrix.python-version }}
-          architecture: x64
-
-      - name: build package
-        run: |
-          python -m pip install -r build_requirements.txt
-          python setup.py bdist_wheel
-
-      - name: Archive dist artifacts
-        uses: actions/upload-artifact@v2
-        with:
-          name: dist-windows-${{ matrix.python-version }}
-          path: dist
-
-  publish-pypi:
-    needs: [build-linux, build-macos-intel, build-macos-arm, build-windows]
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Stage linux 3.7
-        uses: actions/download-artifact@v2
-        with:
-          name: dist-linux-cp37-cp37m
-      - name: Stage linux 3.8
-        uses: actions/download-artifact@v2
-        with:
-          name: dist-linux-cp38-cp38
-      - name: Stage linux 3.9
[…]
