+++ .github/workflows/docs.yml
@@ -1,0 +1,32 @@
+name: docs
+# Based on https://squidfunk.github.io/mkdocs-material/publishing-your-site/
+on:
+  push:
+    branches:
+      - main
+permissions:
+  contents: write
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Configure Git Credentials
+        run: |
+          git config user.name github-actions[bot]
+          git config user.email 41898282+github-actions[bot]@users.noreply.github.com
+      - uses: actions/setup-python@v5
+        with:
+          python-version: 3.12
+      - uses: abatilo/actions-poetry@v2
+        with:
+          poetry-version: 1.8.2
+      - run: echo "cache_id=$(date --utc '+%V')" >> $GITHUB_ENV
+      - uses: actions/cache@v3
+        with:
+          key: mkdocs-material-${{ env.cache_id }}
+          path: .cache
+          restore-keys: |
+            mkdocs-material-
+      - run: poetry install --with docs
+      - run: mkdocs gh-deploy --force
