+++ .github/workflows/docs.yml
@@ -29,4 +29,4 @@ jobs:
           restore-keys: |
             mkdocs-material-
       - run: poetry install --with docs
-      - run: mkdocs gh-deploy --force
+      - run: poetry run mkdocs gh-deploy --force
