+++ README.md
@@ -7,6 +7,7 @@
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10801504.svg)](https://doi.org/10.5281/zenodo.10801504)
 
+- **Documentation**: https://eliotwrobson.github.io/CyNetDiff/
 
 A performance-focused library implementing algorithms for simulating network
 diffusion processes, written in Cython.
