### 2024-03-14 0e33ad6
Made linear threshold much more efficient
