+++ src/cynetdiff/models.pyx
@@ -2,6 +2,7 @@
 from cpython cimport array
 from libcpp.deque cimport deque as cdeque
 from libcpp.unordered_set cimport unordered_set as cset
+from libcpp.algorithm cimport fill
 
 import array
 import random
@@ -140,38 +141,55 @@ cdef class LinearThresholdModel(DiffusionModel):
             self,
             array.array successors,
             array.array successor_starts,
-            array.array predecessors,
-            array.array predecessor_starts,
             *,
             array.array influence = None,
             array.array thresholds = None
         ):
 
+        cdef unsigned int i
+
         self.successors = successors
         self.successor_starts = successor_starts
-        self.predecessors = predecessors
-        self.predecessor_starts = predecessor_starts
-        self.influence = influence
 
-        assert len(self.successor_starts) == len(self.predecessor_starts)
-        # NOTE Assertion below is true because all out-edges must appear as in-edges
-        # somewhere else.
-        assert len(self.predecessors) == len(self.successors)
+        cdef unsigned int n = len(self.successor_starts)
+        cdef unsigned int m = len(self.successors)
+        cdef cvector[unsigned int] in_degrees
 
+        # Setting the influence sent across each edge
+        self.influence.resize(m)
         if influence is not None:
-            assert len(self.predecessors) == len(self.influence)
+            # If provided, copy from user code
+            assert m == len(influence)
+
+            for i in range(m):
+                self.influence[i] = influence[i]
+
+        else:
+            # Otherwise, default to 1/in_degree
+            in_degrees.resize(n)
+            fill(in_degrees.begin(), in_degrees.end(), 0)
+
+            for i in range(m):
+                in_degrees[self.successors[i]] += 1
+
+            for i in range(m):
+                self.influence[i] = 1.0 / in_degrees[self.successors[i]]
 
 
-        self.thresholds.resize(len(self.successor_starts))
-
+        # Setting activation threshold at each node
+        self.thresholds.resize(n)
         if thresholds is not None:
-            assert len(self.successor_starts) == len(thresholds)
+            # If provided, copy from user code
+            assert n == len(thresholds)
 
-            for i in range(len(thresholds)):
+            for i in range(n):
                 self.thresholds[i] = thresholds[i]
         else:
+            # Otherwise, generate
             # Make sure to reassign the size before running this.
             self.reassign_thresholds()
+
+        self.buckets.resize(n)
 
     def set_seeds(self, seeds):
         self.original_seeds.clear()
@@ -189,6 +207,7 @@ cdef class LinearThresholdModel(DiffusionModel):
         self.work_deque.assign(self.original_seeds.begin(), self.original_seeds.end())
         self.seen_set.clear()
         self.seen_set.insert(self.original_seeds.begin(), self.original_seeds.end())
+        fill(self.buckets.begin(), self.buckets.end(), 0.0)
 
     def get_newly_activated_nodes(self):
         for node in self.work_deque:
@@ -200,50 +219,31 @@ cdef class LinearThresholdModel(DiffusionModel):
     # Functions that actually advance the model
     cpdef void advance_until_completion(self):
         while self.work_deque.size() > 0:
-            self.__advance_model(self.work_deque, self.seen_set)
+            self.__advance_model(self.work_deque, self.seen_set, self.buckets)
 
     cpdef void advance_model(self):
-        self.__advance_model(self.work_deque, self.seen_set)
+        self.__advance_model(self.work_deque, self.seen_set, self.buckets)
 
-    cdef inline int __activation_succeeds(self, unsigned int vtx_idx, const cset[unsigned int]& seen_set) except -1 nogil:
-        cdef unsigned int i
-        cdef unsigned int range_end
-        cdef unsigned int parent
+    cdef int __advance_model(
+        self,
[…]
