+++ src/cynetdiff/models.cpp
@@ -1747,17 +1747,17 @@ struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel {
 };
 
 
-/* "cynetdiff/models.pxd":29
+/* "cynetdiff/models.pxd":30
  *     cdef int __advance_model(self, cdeque[unsigned int]& work_deque, cset[unsigned int]& seen_set) except -1 nogil
  * 
  * cdef class LinearThresholdModel(DiffusionModel):             # <<<<<<<<<<<<<<
  *     # Core model parameters
- *     cdef readonly unsigned int[:] successors
+ *     cdef readonly unsigned int[:] starts
  */
 struct __pyx_obj_9cynetdiff_6models_LinearThresholdModel {
   struct __pyx_obj_9cynetdiff_6models_DiffusionModel __pyx_base;
-  __Pyx_memviewslice successors;
-  __Pyx_memviewslice successor_starts;
+  __Pyx_memviewslice starts;
+  __Pyx_memviewslice edges;
   std::vector<float>  influence;
   std::deque<unsigned int>  work_deque;
   std::unordered_set<unsigned int>  seen_set;
@@ -1782,7 +1782,7 @@ struct __pyx_obj_9cynetdiff_6models___pyx_scope_struct__get_newly_activated_node
 };
 
 
-/* "cynetdiff/models.pyx":208
+/* "cynetdiff/models.pyx":212
  *         fill(self.buckets.begin(), self.buckets.end(), 0.0)
  * 
  *     def get_newly_activated_nodes(self):             # <<<<<<<<<<<<<<
@@ -3492,7 +3492,6 @@ static const char __pyx_k_IndexError[] = "IndexError";
 static const char __pyx_k_ValueError[] = "ValueError";
 static const char __pyx_k_pyx_result[] = "__pyx_result";
 static const char __pyx_k_pyx_vtable[] = "__pyx_vtable__";
-static const char __pyx_k_successors[] = "successors";
 static const char __pyx_k_thresholds[] = "thresholds";
 static const char __pyx_k_MemoryError[] = "MemoryError";
 static const char __pyx_k_PickleError[] = "PickleError";
@@ -3518,7 +3517,6 @@ static const char __pyx_k_pyx_PickleError[] = "__pyx_PickleError";
 static const char __pyx_k_setstate_cython[] = "__setstate_cython__";
 static const char __pyx_k_activation_probs[] = "activation_probs";
 static const char __pyx_k_cynetdiff_models[] = "cynetdiff.models";
-static const char __pyx_k_successor_starts[] = "successor_starts";
 static const char __pyx_k_pyx_unpickle_Enum[] = "__pyx_unpickle_Enum";
 static const char __pyx_k_asyncio_coroutines[] = "asyncio.coroutines";
 static const char __pyx_k_cline_in_traceback[] = "cline_in_traceback";
@@ -3648,7 +3646,7 @@ static PyObject *__pyx_pf_9cynetdiff_6models_23IndependentCascadeModel_16activat
 static PyObject *__pyx_pf_9cynetdiff_6models_23IndependentCascadeModel_19_edge_probabilities___get__(struct __pyx_obj_9cynetdiff_6models_IndependentCascadeModel *__pyx_v_self); /* proto */
[…]
