+++ src/cynetdiff/models.cpp
@@ -24958,7 +24958,7 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
  *                 if seen_set.find(child) == seen_set.end():
  *                     child = self.edges[edge_idx]             # <<<<<<<<<<<<<<
  * 
- *                     # TODO remove this assertion once default influence gets fixed.
+ *                     influence = self.influence[edge_idx]
  */
         if (unlikely(!__pyx_v_self->edges.memview)) {PyErr_SetString(PyExc_AttributeError,"Memoryview is not initialized");__PYX_ERR(0, 262, __pyx_L1_error)}
         __pyx_t_12 = __pyx_v_edge_idx;
@@ -24970,24 +24970,24 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
         }
         __pyx_v_child = (*((unsigned int *) ( /* dim=0 */ (__pyx_v_self->edges.data + __pyx_t_12 * __pyx_v_self->edges.strides[0]) )));
 
-        /* "cynetdiff/models.pyx":267
- *                     #assert self.influence is not None
+        /* "cynetdiff/models.pyx":264
+ *                     child = self.edges[edge_idx]
  * 
  *                     influence = self.influence[edge_idx]             # <<<<<<<<<<<<<<
  * 
  *                     # Function is written so that each edge is traversed _once_
  */
-        if (unlikely(!__pyx_v_self->influence.memview)) {PyErr_SetString(PyExc_AttributeError,"Memoryview is not initialized");__PYX_ERR(0, 267, __pyx_L1_error)}
+        if (unlikely(!__pyx_v_self->influence.memview)) {PyErr_SetString(PyExc_AttributeError,"Memoryview is not initialized");__PYX_ERR(0, 264, __pyx_L1_error)}
         __pyx_t_12 = __pyx_v_edge_idx;
         __pyx_t_7 = -1;
         if (unlikely(__pyx_t_12 >= (size_t)__pyx_v_self->influence.shape[0])) __pyx_t_7 = 0;
         if (unlikely(__pyx_t_7 != -1)) {
           __Pyx_RaiseBufferIndexErrorNogil(__pyx_t_7);
-          __PYX_ERR(0, 267, __pyx_L1_error)
+          __PYX_ERR(0, 264, __pyx_L1_error)
         }
         __pyx_v_influence = (*((float *) ( /* dim=0 */ (__pyx_v_self->influence.data + __pyx_t_12 * __pyx_v_self->influence.strides[0]) )));
 
-        /* "cynetdiff/models.pyx":270
+        /* "cynetdiff/models.pyx":267
  * 
  *                     # Function is written so that each edge is traversed _once_
  *                     assert buckets[child] < self.thresholds[child]             # <<<<<<<<<<<<<<
@@ -25004,7 +25004,7 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
                 #endif
                 /*try:*/ {
                   __Pyx_Raise(__pyx_builtin_AssertionError, 0, 0, 0);
-                  __PYX_ERR(0, 270, __pyx_L12_error)
+                  __PYX_ERR(0, 267, __pyx_L12_error)
                 }
                 /*finally:*/ {
                   __pyx_L12_error: {
@@ -25018,10 +25018,10 @@ static int __pyx_f_9cynetdiff_6models_20LinearThresholdModel__LinearThresholdMod
           }
         }
         #else
-        if ((1)); else __PYX_ERR(0, 270, __pyx_L1_error)
+        if ((1)); else __PYX_ERR(0, 267, __pyx_L1_error)
[…]
