+++ pyproject.toml
@@ -1,13 +1,11 @@
 [tool.poetry]
 name = "cynetdiff"
-version = "0.1.6"
+version = "0.1.7"
 description = "A fast network diffusion library"
 authors = ["Eliot W. Robson <eliot.robson24@gmail.com>"]
 license = "MIT"
 readme = "README.md"
 packages = [{ include = "cynetdiff", from = "src" }]
-# TODO try removing the line below and see what happens, may not be needed.
-include = [{ path = "src/cynetdiff/*.so", format = "wheel" }]
 
 #TODO version classifiers may not be needed.
 classifiers = [
