### 2024-03-14 0e9c337
Switch to lazy evaluation
