### 2024-03-15 089727e
Update docs.yml
Change docs to build on new releases
