+++ .github/workflows/docs.yml
@@ -2,8 +2,8 @@ name: docs
 # Based on https://squidfunk.github.io/mkdocs-material/publishing-your-site/
 on:
   push:
-    branches:
-      - main
+    tags:
+      - "v*.*.*"
 permissions:
   contents: write
 jobs:
