+++ pyproject.toml
@@ -30,7 +30,6 @@ documentation = "https://eliotwrobson.github.io/CyNetDiff/"
 
 [tool.poetry.dependencies]
 python = ">=3.9"
-jupyter = "^1.0.0"
 
 [tool.poetry.group.benchmark]
 optional = true
@@ -42,6 +41,7 @@ networkx = "^3.2.1"
 pandas = ">=2.0.0"
 tqdm = "^4.66.2"
 pooch = "^1.8.1"
+jupyter = "^7.1.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.2"
