### 2024-03-23 #165 @isabelizimm
Hello there, thank you so much for opening this issue on the scope of `CyNetDiff` 👋  That is so exciting that you are writing your first Cython package! While we discuss the scope of this package internally, I can answer a few of your questions.

> Are there rules about package maturity before getting a full review?

We ask that any tool/library we review be near "maturing" state, [there is a bit more information in ROpenSci](https://devguide.ropensci.org/authors-guide.html). At submission stage, all major functions should be stable enough to be documented and tested, and the README should make a strong case for the package.

> getting some insight about best practices for the build system would be very helpful.

We would love to have you in [the discourse](https://pyopensci.discourse.group/latest) or Slack (I'll need an email to send an invite)! I personally don't know a lot of knowledge of Cython build systems, but I am certain someone in the community does 😄
