### 2024-03-23 #165 @eliotwrobson
> We ask that any tool/library we review be near "maturing" state, [there is a bit more information in ROpenSci](https://devguide.ropensci.org/authors-guide.html). At submission stage, all major functions should be stable enough to be documented and tested, and the README should make a strong case for the package.

That sounds reasonable! We're gearing up to submit some demo code to a conference, so after that's done and we pick up some more users, I think we'll be able to stabilize the package to this level.

> We would love to have you in [the discourse](https://pyopensci.discourse.group/latest) or Slack (I'll need an email to send an invite)! I personally don't know a lot of knowledge of Cython build systems, but I am certain someone in the community does 😄

@isabelizimm Slack would be great, thank you! My email is eliot.robson24@gmail.com.
