+++ examples/demo_backend.py
@@ -1,6 +1,8 @@
 import random
+import time
 import typing as t
 import warnings
+from functools import wraps
 
 import ndlib.models.epidemics as ep
 import ndlib.models.ModelConfig as mc
@@ -10,6 +12,24 @@ from tqdm.notebook import trange
 
 DiffusionGraphT = t.Union[nx.Graph, nx.DiGraph]
 SeedSetT = set[int]
+
+# Timing decorator
+# from https://stackoverflow.com/a/27737385
+
+
+def timing(f: t.Callable) -> t.Callable:
+    @wraps(f)
+    def wrap(*args, **kw):
+        ts = time.perf_counter()
+        result = f(*args, **kw)
+        te = time.perf_counter()
+        print("func:%r args:[%r, %r] took: %2.4f sec" % (f.__name__, args, kw, te - ts))
+        return result
+
+    return wrap
+
+
+# Pure Python diffusion code
 
 
 def independent_cascade(G: DiffusionGraphT, seeds: SeedSetT) -> list[list[int]]:
@@ -47,6 +67,7 @@ def independent_cascade(G: DiffusionGraphT, seeds: SeedSetT) -> list[list[int]]:
     return res
 
 
+@timing
 def diffuse_python(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> float:
     if not graph.is_directed():
         graph = graph.to_directed()
@@ -59,6 +80,10 @@ def diffuse_python(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) ->
     return res / num_samples
 
 
+# NDlib diffusion wrapper
+
+
+@timing
 def diffuse_ndlib(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> float:
     model = ep.IndependentCascadesModel(graph)
 
@@ -90,6 +115,10 @@ def diffuse_ndlib(graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int) -> 
     return total_infected / num_samples
 
 
+# CyNetDiff diffusion wrapper
+
+
+@timing
 def diffuse_CyNetDiff(
     graph: DiffusionGraphT, seeds: SeedSetT, num_samples: int
 ) -> float:
@@ -104,3 +133,51 @@ def diffuse_CyNetDiff(
         total_activated += model.get_num_activated_nodes()
 
     return total_activated / num_samples
+
+
+# Mini benchmark function
+
+
+def simple_benchmark(
+    graph: DiffusionGraphT,
+    *,
+    num_seeds: int = 10,
+    num_trials: int = 1_000,
+    backends_to_run: t.List[t.Literal["cynetdiff", "ndlib", "python"]] = [
+        "cynetdiff",
+        "ndlib",
+        "python",
+    ],
+) -> None:
+    """
+    A simple benchmark function for three different diffusion
+    implementations.
+    """
+
+    # Number of seed nodes
+    print(f"Number of randomly chosen seed nodes: {num_seeds:_}")
+    print(f"Graph nodes: {graph.number_of_nodes():_}")
+    print(f"Graph edges: {graph.number_of_edges():_}")
+    print(f"Number of trials: {num_trials:_}")
+    seeds = random.sample(list(graph.nodes()), num_seeds)
+
+    if "ndlib" in backends_to_run:
+        print("Starting diffusion with NDlib.")
+        avg_ndlib = diffuse_ndlib(graph, seeds, num_trials)
+
+    if "python" in backends_to_run:
+        print("Starting diffusion with pure Python.")
+        avg_python = diffuse_python(graph, seeds, num_trials)
+
+    if "cynetdiff" in backends_to_run:
[…]
