### 2024-03-25 f2bacc5
Added new get seen nodes function
