### 2024-03-25 50fe60e
Added cool graph visualization.
