### 2024-03-25 9dc4543
Added more plotting example
