+++ examples/demo_backend.py
@@ -1,3 +1,9 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+from __future__ import annotations
+
+import gzip
+import heapq
 import random
 import time
 import typing as t
@@ -10,9 +16,11 @@ import ndlib.models.epidemics as ep
 import ndlib.models.ModelConfig as mc
 import networkx as nx
 import numpy as np
+import pooch
 from coloraide import Color
+from cynetdiff.models import DiffusionModel
 from cynetdiff.utils import networkx_to_ic_model
-from tqdm.notebook import trange
+from tqdm.notebook import tqdm, trange
 
 DiffusionGraphT = t.Union[nx.Graph, nx.DiGraph]
 SeedSetT = set[int]
@@ -381,3 +389,256 @@ def create_plot_for_delta_nodes_infected(
     plt.legend()
     plt.show()
 """
+
+_T = t.TypeVar("_T")
+
+
+class Singleton(type, t.Generic[_T]):
+    """
+    Singleton metaclass, adapted from here:
+    https://stackoverflow.com/a/75308084/2923069
+    """
+
+    _instances: t.Dict[Singleton[_T], _T] = {}
+
+    def __call__(cls, *args: t.Any, **kwargs: t.Any) -> _T:
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+
+class GraphDownloader(metaclass=Singleton):
+    """
+    Registry for all the example data stored on Sariel's site, with corresponding hashes.
+    See: https://sarielhp.org/p/24/frechet_ve/examples/
+    """
+
+    __slots__ = ("__graphs", "__file_fetcher", "__registry")
+
+    __graphs: t.Dict[str, DiffusionGraphT]
+    __file_fetcher: pooch.Pooch
+    __registry: t.Dict[str, str]
+
+    def __init__(self) -> None:
+        self.__registry = {
+            "facebook_combined.txt.gz": "125e84db872eeba443d270c70315c256b0af43a502fcfe51f50621166ad035d7",
+            "twitter_combined.txt.gz": "d9f99b0e6a53b9204b8c215f41b3c10fb99a1e1e783858c012b06d0d3d4bd129",
+            "wiki-Vote.txt.gz": "7d3e53626e14b8b09fb3b396bece9d481ad606bd64ceab066349ff57d4ada7fc",
+            "soc-Epinions1.txt.gz": "69a2dab71fa5e3a0715487599fc16ca17ddc847379325a6c765bbad6e3e36938",
+        }
+
+        self.__file_fetcher = pooch.create(
+            # Use the default cache folder for the operating system
+            path=pooch.os_cache("snap_graphs"),
+            base_url="http://snap.stanford.edu/data/",
+            # The registry specifies the files that can be fetched
+            registry=self.__registry,
+        )
+
+        self.__graphs = {}
+
+    def list_graphs(self) -> t.List[str]:
+        """
+        Get list of graphs that can be returned by this downloader.
+        """
+
+        return list(self.__registry.keys())
+
+    def get_graph(self, name: str) -> DiffusionGraphT:
+        """
+        Get the graph named "name" from this registry.
+        """
+
+        if name not in self.__registry:
+            raise ValueError(f'File with name "{name}" not found in registry.')
+
+        if name not in self.__graphs:
+            graph_data_file = self.__file_fetcher.fetch(name, progressbar=True)
+
+            with gzip.open(graph_data_file, "r") as file:
+                new_graph = nx.read_edgelist(
+                    file,
+                    create_using=nx.Graph(),
+                    nodetype=int,
+                )
+
[…]
