+++ pyproject.toml
@@ -80,7 +80,7 @@ generate-setup-file = false
 python_version = "3.10"
 
 [[tool.mypy.overrides]]
-module = ["networkx.*", "Cython.*", "ndlib.*", "pooch.*"]
+module = ["networkx.*", "Cython.*", "ndlib.*", "pooch.*", "matplotlib.*"]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
