+++ .github/workflows/lint.yml
@@ -29,7 +29,7 @@ jobs:
         with:
           poetry-version: ${{ matrix.poetry-version }}
       - name: Install dependencies
-        run: poetry install --with dev
+        run: poetry install --with dev,benchmark
 
       - name: Run ruff
         run: poetry run ruff check .
