### 2024-03-31 8d27263
Added minor optimization to backend
