### 2024-04-11 #165 @eliotwrobson
@isabelizimm do you have any updates here? We've updated the package substantially since this was opened, and it would be good to have some feedback before we make another pass through and clean things up some more (cc @Batalex since I've seen your comments on other inquiries).
