+++ docs/api/index.md
@@ -1,3 +1,3 @@
 # API
 
-This is the API documentation generated by the class and method docstrings for the `automata` package.
+This is the API documentation generated by the class and method docstrings for the `CyNetDiff` package.
