+++ mkdocs.yml
@@ -107,7 +107,7 @@ extra:
 #   - stylesheets/extra.css
 # extra_javascript:
 #   - javascripts/extra.js
-copyright: Copyright &copy; 2024 Eliot Robson
+copyright: Copyright &copy; 2024 Eliot Robson, Abhishek K. Umrawal
 nav:
   - Introduction:
       - index.md
