### 2024-04-20 #165 @Batalex
Hey @eliotwrobson,
`CynetNetDiff` is in scope for us, would you mind opening a new issue referencing this pre-submission inquiry? Thank you.
